move function from tasks into utils
