+++ corehq/apps/accounting/mixins.py
@@ -3,10 +3,12 @@ from datetime import date, timedelta
 
 from corehq.apps.accounting.models import CreditLine, Subscription
 from corehq.apps.accounting.tasks import (
-    get_unpaid_invoices_over_threshold_by_domain,
     is_subscription_eligible_for_downgrade_process,
 )
-from corehq.apps.accounting.utils import months_from_date
+from corehq.apps.accounting.utils import (
+    months_from_date,
+    get_unpaid_invoices_over_threshold_by_domain,
+)
 from corehq.apps.users.decorators import get_permission_name
 from corehq.apps.users.models import Permissions
 from corehq.util.quickcache import quickcache
