+++ corehq/apps/accounting/tasks.py
@@ -6,7 +6,7 @@ from datetime import date
 
 from django.conf import settings
 from django.db import IntegrityError, transaction
-from django.db.models import F, Q, Sum
+from django.db.models import F, Q
 from django.http import HttpRequest, QueryDict
 from django.template.loader import render_to_string
 from django.utils.translation import ugettext as _
@@ -63,8 +63,9 @@ from corehq.apps.accounting.utils import (
     get_dimagi_from_email,
     log_accounting_error,
     log_accounting_info,
-    get_all_unpaid_saas_invoices,
     get_unpaid_saas_invoices_in_downgrade_daterange,
+    get_unpaid_invoices_over_threshold_by_domain,
+    UNPAID_INVOICE_THRESHOLD,
 )
 from corehq.apps.app_manager.dbaccessors import get_all_apps
 from corehq.apps.domain.models import Domain
@@ -90,8 +91,6 @@ _invoicing_complete_soft_assert = soft_assert(
     ],
     exponential_backoff=False,
 )
-
-UNPAID_INVOICE_THRESHOLD = 100
 
 
 @transaction.atomic
@@ -755,20 +754,6 @@ def _get_domains_with_subscription_invoices_over_threshold(today):
         overdue_invoice, total_overdue_to_date = get_unpaid_invoices_over_threshold_by_domain(today, domain)
         if overdue_invoice:
             yield domain, overdue_invoice, total_overdue_to_date
-
-
-def get_unpaid_invoices_over_threshold_by_domain(today, domain):
-    for overdue_invoice in get_unpaid_saas_invoices_in_downgrade_daterange(today).filter(
-        subscription__subscriber__domain=domain
-    ):
-        total_overdue_by_domain_and_invoice_date = get_all_unpaid_saas_invoices().filter(
-            Q(date_due__lte=overdue_invoice.date_due)
-            | (Q(date_due__isnull=True) & Q(date_end__lte=overdue_invoice.date_end)),
-            subscription__subscriber__domain=domain,
-        ).aggregate(Sum('balance'))['balance__sum']
-        if total_overdue_by_domain_and_invoice_date >= UNPAID_INVOICE_THRESHOLD:
-            return overdue_invoice, total_overdue_by_domain_and_invoice_date
-    return None, None
 
 
 def get_accounts_with_customer_invoices_over_threshold(today):
