+++ corehq/apps/accounting/utils.py
@@ -6,6 +6,7 @@ from django.conf import settings
 from django.template.loader import render_to_string
 from django.urls import reverse
 from django.utils.translation import ugettext_lazy as _
+from django.db.models import Q, Sum
 
 from django_prbac.models import Grant, Role, UserRole
 
@@ -25,6 +26,7 @@ from corehq.util.view_utils import absolute_reverse
 logger = logging.getLogger('accounting')
 
 EXCHANGE_RATE_DECIMAL_PLACES = 9
+UNPAID_INVOICE_THRESHOLD = 100
 
 
 def log_accounting_error(message, show_stack_trace=False):
@@ -60,6 +62,20 @@ def get_unpaid_saas_invoices_in_downgrade_daterange(today):
     return get_all_unpaid_saas_invoices().filter(
         date_due__lte=today - datetime.timedelta(days=1)
     ).order_by('date_due').select_related('subscription__subscriber')
+
+
+def get_unpaid_invoices_over_threshold_by_domain(today, domain):
+    for overdue_invoice in get_unpaid_saas_invoices_in_downgrade_daterange(today).filter(
+        subscription__subscriber__domain=domain
+    ):
+        total_overdue_by_domain_and_invoice_date = get_all_unpaid_saas_invoices().filter(
+            Q(date_due__lte=overdue_invoice.date_due)
+            | (Q(date_due__isnull=True) & Q(date_end__lte=overdue_invoice.date_end)),
+            subscription__subscriber__domain=domain,
+        ).aggregate(Sum('balance'))['balance__sum']
+        if total_overdue_by_domain_and_invoice_date >= UNPAID_INVOICE_THRESHOLD:
+            return overdue_invoice, total_overdue_by_domain_and_invoice_date
+    return None, None
 
 
 def fmt_feature_rate_dict(feature, feature_rate=None):
